+++ test/JDBC/expected/xml_value-vu-verify.out
@@ -1828,3 +1828,41 @@ GO
 ~~ERROR (Message: The value function requires 2 argument(s).)~~
 
 
+-- check for the special tag not being used in data or query
+DECLARE @x XML = '<magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>test</magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>  ';
+SELECT @x.value('(/*)', 'NVARCHAR(100)') AS c;
+go
+~~START~~
+nvarchar
+~~ERROR (Code: 33557097)~~
+
+~~ERROR (Message: Babelfish-internal XML tag cannot be used in XML data: [<magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>])~~
+
+
+DECLARE @x XML = '<MAGIC_BBF_XMLNODES_945193483C854AF5A887B50698B99B05_TAG>test</MAGIC_BBF_XMLNODES_945193483C854AF5A887B50698B99B05_TAG>  ';
+SELECT @x.value('(/*)', 'NVARCHAR(100)') AS c;
+go
+~~START~~
+nvarchar
+test
+~~END~~
+
+
+DECLARE @x XML = '<magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>test</magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>  ';
+SELECT @x.value('(/magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag)', 'NVARCHAR(100)') AS c;
+go
+~~START~~
+nvarchar
+~~ERROR (Code: 33557097)~~
+
+~~ERROR (Message: Babelfish-internal XML tag cannot be used in XML data: [<magic_bbf_xmlnodes_945193483c854af5a887b50698b99b05_tag>])~~
+
+
+DECLARE @x XML = '<MAGIC_BBF_XMLNODES_945193483C854AF5A887B50698B99B05_TAG>test</MAGIC_BBF_XMLNODES_945193483C854AF5A887B50698B99B05_TAG>  ';
+SELECT @x.value('(/MAGIC_BBF_XMLNODES_945193483C854AF5A887B50698B99B05_TAG)', 'NVARCHAR(100)') AS c;
+go
+~~START~~
+nvarchar
+test
+~~END~~
+
